# server/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CleanupExecution": {
        "properties": {
          "confirmation_id": {
            "title": "Confirmation Id",
            "type": "string"
          }
        },
        "required": [
          "confirmation_id"
        ],
        "title": "CleanupExecution",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AndroidAPIKey": {
        "scheme": "bearer",
        "type": "http"
      },
      "WebSession": {
        "in": "cookie",
        "name": "dokura_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Dokura",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/cache-cleanup/execute": {
      "post": {
        "operationId": "cleanup_execute_api_v1_admin_cache_cleanup_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanupExecution"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Cleanup Execute Api V1 Admin Cache Cleanup Execute Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WebSession": []
          }
        ],
        "summary": "Cleanup Execute",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/v1/admin/cache-cleanup/preview": {
      "post": {
        "operationId": "cleanup_preview_api_v1_admin_cache_cleanup_preview_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cleanup Preview Api V1 Admin Cache Cleanup Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WebSession": []
          }
        ],
        "summary": "Cleanup Preview",
        "tags": [
          "缓存管理"
        ]
      }
    },
...
    "/api/v1/admin/scan": {
      "get": {
        "operationId": "scan_status_api_v1_admin_scan_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scan Status Api V1 Admin Scan Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WebSession": []
          }
        ],
        "summary": "Scan Status",
        "tags": [
          "扫描与后台任务"
        ]
      },
      "post": {
        "operationId": "request_scan_api_v1_admin_scan_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Request Scan Api V1 Admin Scan Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WebSession": []
          }
        ],
        "summary": "Request Scan",
        "tags": [
          "扫描与后台任务"
        ]
      }
    },
    "/api/v1/admin/tasks": {
      "get": {
        "operationId": "task_status_api_v1_admin_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Task Status Api V1 Admin Tasks Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WebSession": []
          }
        ],
        "summary": "Task Status",
        "tags": [
          "扫描与后台任务"
        ]
      }
    },
...
  }
}